// supabase/db-types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  public: {
    Tables: {
      candidate_aspirations: {
        Row: {
          aspiration_summary: string | null
          candidate_id: string | null
          created_at: string
          embedding: string | null
          gps: unknown | null
          id: string
          last_proompted_on: string | null
          mission_statement: string
          period: string
          updated_at: string | null
          vibe_check: Json | null
          vying_for: Database["public"]["Enums"]["positions"]
          vying_in: string | null
        }
        Insert: {
          aspiration_summary?: string | null
          candidate_id?: string | null
          created_at?: string
          embedding?: string | null
          gps?: unknown | null
          id?: string
          last_proompted_on?: string | null
          mission_statement: string
          period: string
          updated_at?: string | null
          vibe_check?: Json | null
          vying_for: Database["public"]["Enums"]["positions"]
          vying_in?: string | null
        }
        Update: {
          aspiration_summary?: string | null
          candidate_id?: string | null
          created_at?: string
          embedding?: string | null
          gps?: unknown | null
          id?: string
          last_proompted_on?: string | null
          mission_statement?: string
          period?: string
          updated_at?: string | null
          vibe_check?: Json | null
          vying_for?: Database["public"]["Enums"]["positions"]
          vying_in?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "candidate_aspirations_candidate_id_fkey"
            columns: ["candidate_id"]
            isOneToOne: false
            referencedRelation: "candidates"
            referencedColumns: ["id"]
          },
        ]
      }
      candidate_reviews: {
        Row: {
          candidate_id: string | null
          comment: string | null
          created_at: string
          id: number
          rating: number | null
          reviewr_id: string | null
        }
        Insert: {
          candidate_id?: string | null
          comment?: string | null
          created_at?: string
          id?: number
          rating?: number | null
          reviewr_id?: string | null
        }
        Update: {
          candidate_id?: string | null
          comment?: string | null
          created_at?: string
          id?: number
          rating?: number | null
          reviewr_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "candidate_reviews_candidate_id_fkey"
            columns: ["candidate_id"]
            isOneToOne: false
            referencedRelation: "candidates"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "candidate_reviews_review_by_fkey"
            columns: ["reviewr_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
        ]
      }
      candidates: {
        Row: {
          account_id: string
          avatar_url: string | null
          banner_url: string | null
          bio: string | null
          candidate_summary: string | null
          created_at: string
          embedding: string | null
          id: string
          last_proompted_on: string | null
          name: string
          updated_at: string | null
          vibe_check: Json | null
        }
        Insert: {
          account_id?: string
          avatar_url?: string | null
          banner_url?: string | null
          bio?: string | null
          candidate_summary?: string | null
          created_at?: string
          embedding?: string | null
          id?: string
          last_proompted_on?: string | null
          name: string
          updated_at?: string | null
          vibe_check?: Json | null
        }
        Update: {
          account_id?: string
          avatar_url?: string | null
          banner_url?: string | null
          bio?: string | null
          candidate_summary?: string | null
          created_at?: string
          embedding?: string | null
          id?: string
          last_proompted_on?: string | null
          name?: string
          updated_at?: string | null
          vibe_check?: Json | null
        }
        Relationships: [
          {
            foreignKeyName: "candidates_account_id_fkey"
            columns: ["account_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "candidates_id_fkey"
            columns: ["id"]
            isOneToOne: true
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]
// ...
export type Enums<
  PublicEnumNameOrOptions extends
    | keyof PublicSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema["Enums"]
    ? PublicSchema["Enums"][PublicEnumNameOrOptions]
    : never